Refactor advanced query parsing
